feat: move routing from App to RouterApp, rename RoutsList
